requirements.txt
keras
numpy
cv2
pillow
matplotlib
graphviz
pydot
pyqt5
h5py